% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest.R
\name{ver_latest}
\alias{ver_latest}
\title{Find the latest version code}
\usage{
ver_latest(x)
}
\arguments{
\item{x}{\code{\link[=character]{character()}}\cr
A vector of version codes that start with a number (as in "1.5.0"), i.e.
without initial "v" (as in "v1.5.0").}
}
\value{
A character vector with one element.
}
\description{
\code{ver_latest()} returns the latest of the version codes.
}
\examples{
ver_latest(c("1.3-0", "1.4-1", "0.0.0.9000", "1.4-0a", "1.4-0"))

}
\seealso{
\code{\link{ver_oldest}()}
}
