% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versionsort.R
\name{ver_sort}
\alias{ver_sort}
\title{Sort version codes}
\usage{
ver_sort(x)
}
\arguments{
\item{x}{\code{\link[=character]{character()}}\cr
A vector of version codes that start with a number (as in "1.5.0"), i.e.
without initial "v" (as in "v1.5.0").}
}
\value{
A character vector containing the same elements as input, but
reordered.
}
\description{
\code{ver_sort()} returns a sorted vector of version codes, where
sorting is done alphanumerically.
}
\examples{
ver_sort(c("1.5-0", "1.4-1", "0.0.0.9000", "1.4-0a", "1.4-0"))

}
\seealso{
\code{\link{ver_order}()}
}
